move JavaScriptPropTypes to own dir
