+++ packages/quicktype-core/src/language/JavaScriptPropTypes/language.ts
@@ -1,0 +1,47 @@
+import { type RenderContext } from "../../Renderer";
+import { EnumOption, type Option, getOptionValues } from "../../RendererOptions";
+import { AcronymStyleOptions, acronymOption } from "../../support/Acronyms";
+import { convertersOption } from "../../support/Converters";
+import { TargetLanguage } from "../../TargetLanguage";
+import { type FixMeOptionsAnyType, type FixMeOptionsType } from "../../types";
+
+import { JavaScriptPropTypesRenderer } from "./JavaScriptPropTypesRenderer";
+
+export const javaScriptPropTypesOptions = {
+    acronymStyle: acronymOption(AcronymStyleOptions.Pascal),
+    converters: convertersOption(),
+    moduleSystem: new EnumOption(
+        "module-system",
+        "Which module system to use",
+        [
+            ["common-js", false],
+            ["es6", true]
+        ],
+        "es6"
+    )
+};
+
+export class JavaScriptPropTypesTargetLanguage extends TargetLanguage {
+    protected getOptions(): Array<Option<FixMeOptionsAnyType>> {
+        return [javaScriptPropTypesOptions.acronymStyle, javaScriptPropTypesOptions.converters];
+    }
+
+    public constructor(
+        displayName = "JavaScript PropTypes",
+        names: string[] = ["javascript-prop-types"],
+        extension = "js"
+    ) {
+        super(displayName, names, extension);
+    }
+
+    protected makeRenderer(
+        renderContext: RenderContext,
+        untypedOptionValues: FixMeOptionsType
+    ): JavaScriptPropTypesRenderer {
+        return new JavaScriptPropTypesRenderer(
+            this,
+            renderContext,
+            getOptionValues(javaScriptPropTypesOptions, untypedOptionValues)
+        );
+    }
+}
